Move Filesystem::OsSysCalls to Api::OsSysCalls.
This makes it appopriate to use the OsSysCalls abstraction
in other areas of the code-base, such as hot-restart.
