+++ include/envoy/filesystem/filesystem.h
@@ -10,31 +10,6 @@
 namespace Envoy {
 namespace Filesystem {
 
-class OsSysCalls {
-public:
-  virtual ~OsSysCalls(){};
-
-  /**
-   * Open file by full_path with given flags and mode.
-   * @return file descriptor.
-   */
-  virtual int open(const std::string& full_path, int flags, int mode) PURE;
-
-  /**
-   * Write num_bytes to fd from buffer.
-   * @return number of bytes written if non negative, otherwise error code.
-   */
-  virtual ssize_t write(int fd, const void* buffer, size_t num_bytes) PURE;
-
-  /**
-   * Release all resources allocated for fd.
-   * @return zero on success, -1 returned otherwise.
-   */
-  virtual int close(int fd) PURE;
-};
-
-typedef std::unique_ptr<OsSysCalls> OsSysCallsPtr;
-
 /**
  * Abstraction for a file on disk.
  */
