+++ source/common/api/BUILD
@@ -14,7 +14,17 @@ envoy_cc_library(
     hdrs = ["api_impl.h"],
     deps = [
         "//include/envoy/api:api_interface",
+        "//source/common/api:os_sys_calls_lib",
         "//source/common/event:dispatcher_lib",
         "//source/common/filesystem:filesystem_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "os_sys_calls_lib",
+    srcs = ["os_sys_calls_impl.cc"],
+    hdrs = ["os_sys_calls_impl.h"],
+    deps = [
+        "//include/envoy/api:os_sys_calls_interface",
     ],
 )
