+++ source/common/api/api_impl.cc
@@ -3,6 +3,7 @@
 #include <chrono>
 #include <string>
 
+#include "common/api/os_sys_calls_impl.h"
 #include "common/event/dispatcher_impl.h"
 #include "common/filesystem/filesystem_impl.h"
 
@@ -14,8 +15,7 @@ Event::DispatcherPtr Impl::allocateDispatcher() {
 }
 
 Impl::Impl(std::chrono::milliseconds file_flush_interval_msec)
-    : os_sys_calls_(new Filesystem::OsSysCallsImpl()),
-      file_flush_interval_msec_(file_flush_interval_msec) {}
+    : os_sys_calls_(new OsSysCallsImpl()), file_flush_interval_msec_(file_flush_interval_msec) {}
 
 Filesystem::FileSharedPtr Impl::createFile(const std::string& path, Event::Dispatcher& dispatcher,
                                            Thread::BasicLockable& lock, Stats::Store& stats_store) {
