+++ test/mocks/filesystem/mocks.h
@@ -13,27 +13,6 @@
 namespace Envoy {
 namespace Filesystem {
 
-class MockOsSysCalls : public OsSysCalls {
-public:
-  MockOsSysCalls();
-  ~MockOsSysCalls();
-
-  // Filesystem::OsSysCalls
-  ssize_t write(int fd, const void* buffer, size_t num_bytes) override;
-  int open(const std::string& full_path, int flags, int mode) override;
-  MOCK_METHOD1(close, int(int));
-
-  MOCK_METHOD3(open_, int(const std::string& full_path, int flags, int mode));
-  MOCK_METHOD3(write_, ssize_t(int, const void*, size_t));
-
-  size_t num_writes_;
-  size_t num_open_;
-  Thread::MutexBasicLockable write_mutex_;
-  Thread::MutexBasicLockable open_mutex_;
-  std::condition_variable_any write_event_;
-  std::condition_variable_any open_event_;
-};
-
 class MockFile : public File {
 public:
   MockFile();
